backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-learn==1.1.1
Werkzeug==2.1.2